# api/openapi.yaml
openapi: 3.0.0
info:
  title: Vector Search Service API
  description: |
    A vector search service providing semantic search capabilities using PostgreSQL + pgvector.
    
    This service provides document ingestion, embedding generation, and similarity search operations
    designed for integration with Elyra pipelines and other AI workflows.
    
    ## Features
    - Document ingestion with automatic chunking
    - Semantic similarity search
    - Batch processing for pipeline integration
    - Collection management for document organization
    - Asynchronous job tracking
    
    ## Authentication
    This service uses JWT tokens for authentication. Include the token in the Authorization header:
    `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: Vector Search Service
    url: https://github.com/vector-search-service
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://vector-search-service.openshift.apps.example.com
    description: OpenShift deployment
paths:
...
  /api/v1/collections:
    post:
      summary: Create collection
      description: Create a new collection for organizing documents
      tags:
        - collections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
      responses:
        '200':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionCreateResponse'
        '409':
          description: Collection already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/collections/{collection_id}:
    get:
      summary: Get collection information
      description: Retrieve detailed information about a specific collection
      tags:
        - collections
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Unique identifier of the collection
          schema:
            type: string
      responses:
        '200':
          description: Collection information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionInfo'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete collection
      description: Delete a collection and optionally all its documents
      tags:
        - collections
      parameters:
        - name: collection_id
          in: path
          required: true
          description: Unique identifier of the collection to delete
          schema:
            type: string
        - name: force
          in: query
          required: false
          description: Force delete collection even if it contains documents
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Collection deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  collection_id:
                    type: string
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  force_delete:
                    type: boolean
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Collection contains documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CollectionCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the collection
          example: "Technical Documentation"
        description:
          type: string
          description: Description of the collection
          example: "Collection for technical documentation and guides"
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for the collection
    CollectionInfo:
      type: object
      required:
        - id
        - name
        - document_count
        - embedding_count
        - created_at
        - updated_at
        - metadata
      properties:
        id:
          type: string
          description: Unique identifier of the collection
        name:
          type: string
          description: Name of the collection
        description:
          type: string
          description: Description of the collection
        document_count:
          type: integer
          description: Number of documents in the collection
        embedding_count:
          type: integer
          description: Number of embeddings in the collection
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        metadata:
          type: object
          additionalProperties: true
          description: Collection metadata
    CollectionCreateResponse:
      type: object
      required:
        - collection_id
        - name
        - status
        - created_at
      properties:
        collection_id:
          type: string
          description: Unique identifier of the created collection
        name:
          type: string
          description: Name of the collection
        status:
          type: string
          enum: [created, failed]
          description: Status of the collection creation
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        request_id:
          type: string
          description: Request ID for debugging
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []
...