# src/api/collections.py
from fastapi import APIRouter, HTTPException, status
from typing import Dict, Any
import uuid
import logging
from datetime import datetime

from .models import (
    CollectionCreateRequest,
    CollectionCreateResponse,
    CollectionInfo,
    ErrorResponse
)

logger = logging.getLogger(__name__)
router = APIRouter()


@router.post("/collections",
             response_model=CollectionCreateResponse,
             responses={
                 422: {"model": ErrorResponse, "description": "Validation error"},
                 409: {"model": ErrorResponse, "description": "Collection already exists"},
                 500: {"model": ErrorResponse, "description": "Internal server error"}
             })
async def create_collection(request: CollectionCreateRequest) -> CollectionCreateResponse:
    """
    Create a new collection for organizing documents.
    
    Collections provide a way to group related documents together
    and perform searches within specific document sets.
    """
    try:
        logger.info(f"Creating new collection: {request.name}")
        
        # Generate unique collection ID
        collection_id = str(uuid.uuid4())
        
        # TODO: Implement actual collection creation logic
        # This would:
        # 1. Check if collection name already exists
        # 2. Create collection in database
        # 3. Set up collection metadata
        # 4. Initialize collection structure
        
        created_at = datetime.now()
        
        return CollectionCreateResponse(
            collection_id=collection_id,
            name=request.name,
            status="created",
            created_at=created_at
        )
        
    except Exception as e:
        logger.error(f"Collection creation failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Collection creation failed: {str(e)}"
        )


@router.get("/collections/{collection_id}",
            response_model=CollectionInfo,
            responses={
                404: {"model": ErrorResponse, "description": "Collection not found"},
                500: {"model": ErrorResponse, "description": "Internal server error"}
            })
async def get_collection(collection_id: str) -> CollectionInfo:
    """
    Retrieve detailed information about a specific collection.
    
    Returns metadata, document counts, and other collection statistics.
    """
    try:
        logger.info(f"Retrieving collection: {collection_id}")
        
        # TODO: Implement actual collection retrieval logic
        # For now, return mock data
        
        # Mock collection info
        mock_collection = CollectionInfo(
            id=collection_id,
            name="Sample Collection",
            description="A sample collection for demonstration",
            document_count=15,
            embedding_count=90,
            created_at=datetime.now(),
            updated_at=datetime.now(),
            metadata={"type": "sample", "version": "1.0"}
        )
        
        return mock_collection
        
    except Exception as e:
        logger.error(f"Collection retrieval failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Collection retrieval failed: {str(e)}"
        )


@router.delete("/collections/{collection_id}",
               responses={
                   200: {"description": "Collection deleted successfully"},
                   404: {"model": ErrorResponse, "description": "Collection not found"},
                   409: {"model": ErrorResponse, "description": "Collection contains documents"},
                   500: {"model": ErrorResponse, "description": "Internal server error"}
               })
async def delete_collection(collection_id: str, force: bool = False) -> Dict[str, Any]:
    """
    Delete a collection and optionally all its documents.
    
    By default, collections with documents cannot be deleted.
    Use force=true to delete collection and all its documents.
    """
    try:
        logger.info(f"Deleting collection: {collection_id} (force={force})")
        
        # TODO: Implement actual collection deletion logic
        # This would:
        # 1. Check if collection exists
        # 2. Check if collection has documents (unless force=true)
        # 3. Delete all documents and embeddings if force=true
        # 4. Delete collection metadata
        # 5. Clean up collection structure
        
        return {
            "message": f"Collection {collection_id} deleted successfully",
            "collection_id": collection_id,
            "status": "deleted",
            "timestamp": datetime.now().isoformat(),
            "force_delete": force
        }
        
    except Exception as e:
        logger.error(f"Collection deletion failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Collection deletion failed: {str(e)}"
        )